+++ CHANGELOG.rst
@@ -1,3 +1,13 @@
+Unreleased
+----------
+
+Changed
+ * Rename ``ThermostatDevice`` and ``ThermostatAttributes`` into ``HvacDevice``
+   and ``HvacAttributes``, respectively. This is in anticipation that there are
+   other HVAC products with similar API.
+ * Rename ``DeviceType.DEVICE`` into ``DeviceType.HVAC`` to be more descriptive
+   and not reserve the most general name for just one kind of device.
+
 Version 0.1
 -----------